se solucionan errores con los modales de edicion, props
