+++ src/hooks/useAxiosPrivate.jsx
@@ -1,11 +1,14 @@
-import { useEffect } from 'react';
+import { useContext, useEffect } from 'react';
 import { useAuth } from './useAuth';
 import { privateBackendAPI } from '../api/backendAPI';
 import { useNavigate } from 'react-router-dom';
+import { ToastContext } from '../context/ToastContext';
 
 export const useAxiosPrivate = () => {
 	const { auth } = useAuth();
 	const navigate = useNavigate();
+	const { addToast } = useContext(ToastContext);
+
 	useEffect(() => {
 		const requestIntercept = privateBackendAPI.interceptors.request.use(
 			(config) => {
@@ -22,10 +25,15 @@ export const useAxiosPrivate = () => {
 				return response;
 			},
 			(error) => {
-				if (error.response.status === 401) {
+				if (error?.response?.status === 401) {
 					// Redireccionar al no tener autenticación
 					navigate('/login');
 				}
+				addToast({
+					message: 'Error en la operación - ' + error,
+					variant: 'error',
+				});
+				
 			}
 		);
